feat: add period update and cancellation in mypage

+++ backend/src/main/java/com/swimpulse/mypage/MyPageMetrics.java
@@ -4,7 +4,7 @@ public record MyPageMetrics(
 		int subscriptionCount,
 		long upcomingSubscriptionCount,
 		long openSubscriptionCount,
-		int notificationCount,
+		long notificationCount,
 		long unreadNotificationCount,
 		long activeDeviceCount
 ) {
